:recycle: extract getVersionNumber() function to services
