+++ api/src/certification/configuration/domain/services/get-version-number.js
@@ -1,0 +1,11 @@
+export function getVersionNumber(date = new Date()) {
+  const pad = (n) => String(n).padStart(2, '0');
+  return (
+    date.getUTCFullYear().toString() +
+    pad(date.getUTCMonth() + 1) +
+    pad(date.getUTCDate()) +
+    pad(date.getUTCHours()) +
+    pad(date.getUTCMinutes()) +
+    pad(date.getSeconds())
+  );
+}
